Byte Track and Color Filter
The project was set up on DeepSORT before this update. It was a little bit complicated because it has many parameter to play with. This was time consuming for optimization. That's why I could not find perfect fit parameters and this brought huge error rates on track_ids.

After all these problems, I decided to use Byte Track on the project. It simply works with YOLO detection model. Its results are way better than our Deep SORT model.

The other thing is that color filter. I have did wrong update on "out.write(frame)" statement. Its position was wrong. Now, it is fixed.