matting.py debug for overwritten humans
In this update, humans do not block other humans when they walk together. Before this update, one of the humans was parsed on another human when they near to each other.

I have changed the black frames to 4 dimension. Alpha is used because it makes the black area of the processed images transparent. This enables us to parse humans together.

Also, comments are added to the code. New functions are added to debug the issue.